[REFACTOR] Top Nav reorganized
