+++ src/components/TopNav/index.js
@@ -1,3 +1,3 @@
-import TopNav from "./TopNav";
+import TopNav from "./top-nav";
 
 export default TopNav;
